requirements.txt
gradio
Pillow
transformers
gtts
torch
sentencepiece